+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+## Fixed
+
+- Fixed failing to create switcher due to failure while deserializaing komorebi monitor information
+
 ## [0.7.0] - 2025-05-12
 
 ## Added